// types/inventoryVoucher/SalesVoucherTypes.d.ts
export interface SalesVoucherFormData {
  partyName: string;
  invoiceDate: Date;
  partyId: string;
  gstChoice: string;
  gstAmount: string;
  bankId: string;
  transMode: string;
}
// ...
export interface SalesVoucherTableData {
  Id: number;
  Order_Date: string;
  Order_No: string;
  Party_Name: string;
  Party_Id: number;
  Total_Order: string;
  Order_Status: string;
  DesignRow: {
    Design_Id: number;
    Design_Name: string;
    Design_No: string;
    Order_Qnty: string;
    Design_Rate: string;
    Wt: string;
    Wt_Rate: string;
    Tot_Wt: string;
    Polish: string;
    Tot_Polish: string;
    Image: string;
    ItemRow: {
      Item_Id: number;
      Item_Name: string;
      Item_Qnty: string;
      Item_Rate: string;
      Making_Rate: string;
      Item_Tot: string;
    }[];
  }[];
}